+++ test-projects/free-app/src/integration-test/groovy/freeApp/FreeMarkerTemplateServiceTests.groovy
@@ -1,5 +1,4 @@
-package freeApp
-
+import grails.plugin.freemarker.FreeMarkerTemplateService
 import grails.test.mixin.integration.Integration
 import spock.lang.Specification
