+++ test-projects/free-app/src/integration-test/groovy/freeApp/GrailsFreeMarkerViewResolverTests.groovy
@@ -1,4 +1,4 @@
-package freeApp
+
 
 import grails.test.mixin.integration.Integration
 import spock.lang.Specification
