+++ test-projects/free-app/src/integration-test/groovy/freeApp/UserDefinedTagLibTests.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package freeApp
+
 
 import freemarker.template.Configuration
 import freemarker.template.Template
